+++ src/App/src/Entity/UserMessageEntity.php
@@ -1,0 +1,112 @@
+<?php
+/**
+ * @see https://github.com/dotkernel/frontend/ for the canonical source repository
+ * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
+ * @license https://github.com/dotkernel/frontend/blob/master/LICENSE.md MIT License
+ */
+
+declare(strict_types=1);
+
+namespace Frontend\App\Entity;
+
+/**
+ * Class UserMessageEntity
+ * @package Frontend\App\Entity
+ */
+class UserMessageEntity
+{
+    /** @var  int */
+    protected $id;
+
+    /** @var  string */
+    protected $email;
+
+    /** @var  string */
+    protected $name;
+
+    /** @var  string */
+    protected $subject;
+
+    /** @var  string */
+    protected $message;
+
+    /**
+     * @return int
+     */
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    /**
+     * @param int $id
+     */
+    public function setId($id)
+    {
+        $this->id = $id;
+    }
+
+    /**
+     * @return string
+     */
+    public function getEmail(): string
+    {
+        return $this->email ?? '';
+    }
+
+    /**
+     * @param string $email
+     */
+    public function setEmail(string $email)
+    {
+        $this->email = $email;
+    }
+
+    /**
+     * @return string
+     */
+    public function getName(): string
+    {
+        return $this->name ?? '';
+    }
+
+    /**
+     * @param string $name
+     */
+    public function setName(string $name)
+    {
+        $this->name = $name;
+    }
+
+    /**
+     * @return string
+     */
+    public function getSubject(): string
+    {
+        return $this->subject ?? '';
+    }
+
+    /**
+     * @param string $subject
+     */
+    public function setSubject(string $subject)
+    {
+        $this->subject = $subject;
+    }
+
+    /**
+     * @return string
+     */
+    public function getMessage(): string
+    {
+        return $this->message ?? '';
+    }
+
+    /**
+     * @param string $message
+     */
+    public function setMessage(string $message)
+    {
+        $this->message = $message;
+    }
+}