+++ src/User/src/Listener/UserEventsListener.php
@@ -11,6 +11,7 @@ namespace Frontend\User\Listener;
 
 use Dot\AnnotatedServices\Annotation\Inject;
 use Dot\AnnotatedServices\Annotation\Service;
+use Dot\FlashMessenger\FlashMessengerInterface;
 use Dot\User\Entity\ConfirmTokenEntity;
 use Dot\User\Entity\ResetTokenEntity;
 use Dot\User\Event\TokenEvent;
@@ -44,14 +45,19 @@ class UserEventsListener implements UserEventListenerInterface, TokenEventListen
     /** @var  UserMailerService */
     protected $userMailer;
 
+    /** @var FlashMessengerInterface  */
+    protected $messenger;
+
     /**
      * UserEventsListener constructor.
      * @param UserMailerService $userMailer
+     * @param FlashMessengerInterface $flashMessenger
      *
-     * @Inject({UserMailerService::class})
+     * @Inject({UserMailerService::class, FlashMessengerInterface::class})
      */
-    public function __construct(UserMailerService $userMailer)
+    public function __construct(UserMailerService $userMailer, FlashMessengerInterface $flashMessenger)
     {
+        $this->messenger = $flashMessenger;
         $this->userMailer = $userMailer;
     }
 
@@ -65,6 +71,8 @@ class UserEventsListener implements UserEventListenerInterface, TokenEventListen
         if ($token instanceof ConfirmTokenEntity) {
             /** @var UserEntity $user */
             $user = $e->getParam('user');
+            // we silently fail to send the e-mail, probably the service was not configured
+            // the failure should be logged instead
             $this->userMailer->sendActivationEmail($user, $token);
         }
     }
@@ -78,6 +86,8 @@ class UserEventsListener implements UserEventListenerInterface, TokenEventListen
         if ($token instanceof ResetTokenEntity) {
             /** @var UserEntity $user */
             $user = $e->getParam('user');
+            // we silently fail to send the e-mail, probably the service was not configured
+            // the failure should be logged instead
             $this->userMailer->sendPasswordRecoveryEmail($user, $token);
         }
     }